postgres dev module

+++ environments/dev/modules/postgres/manifests/install.pp
@@ -8,17 +8,17 @@ class postgres::install {
 	file { '/db/pgsql/data': mode => 0700, owner => postgres, group => postgres, require => File['/db/pgsql'] }
 	file { '/db/log': owner => postgres, group => postgres, require => File['/db'] }
   file { '/db/log/pg_xlog': owner => postgres, group => postgres, require => File['/db/log'] }
-	
-	notify { 'init_message': message => "postgres database not yet initialized; not placing config files.", unless => $pgsql_version }
+
+	unless $pgsql_version { notify { 'postgres database not yet initialized; not placing config files.' } }
   package { "$::postgres::packages": ensure => installed, allow_virtual => true }
 #
 #	package { "$::postgres::packages": ensure => installed, allow_virtual => false }
 # 
+  unless $pgsql_version {
     exec { "delete_default_configs":
 		  command => "/bin/rm -f /db/pgsql/data/{postgresql.conf,pg_hba.conf}",
-		  unless  => $pgsql_version,
 		  require => Package['$::postgres::packages']
     }
-
+  }
     file { "/etc/sysconfig/pgsql/$::postgres::service": mode => 0644, source => "$source/conf" }
 }
